+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/apps/AppResourceSecretSecretPermission.java
@@ -1,0 +1,15 @@
+// Code generated from OpenAPI specs by Databricks SDK Generator. DO NOT EDIT.
+
+package com.databricks.sdk.service.apps;
+
+import com.databricks.sdk.support.Generated;
+
+/**
+ * Permission to grant on the secret scope. Supported permissions are: "READ", "WRITE", "MANAGE".
+ */
+@Generated
+public enum AppResourceSecretSecretPermission {
+  MANAGE,
+  READ,
+  WRITE,
+}