+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/catalog/ConnectionType.java
@@ -11,6 +11,7 @@ public enum ConnectionType {
   DATABRICKS,
   GLUE,
   HIVE_METASTORE,
+  HTTP,
   MYSQL,
   POSTGRESQL,
   REDSHIFT,