+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/jobs/RunParameters.java
@@ -51,7 +51,7 @@ public class RunParameters {
   @JsonProperty("notebook_params")
   private Map<String, String> notebookParams;
 
-  /** */
+  /** Controls whether the pipeline should perform a full refresh */
   @JsonProperty("pipeline_params")
   private PipelineParams pipelineParams;