+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/sql/CreateWarehouseRequest.java
@@ -13,7 +13,8 @@ public class CreateWarehouseRequest {
    * The amount of time in minutes that a SQL warehouse must be idle (i.e., no RUNNING queries)
    * before it is automatically stopped.
    *
-   * <p>Supported values: - Must be == 0 or >= 10 mins - 0 indicates no autostop.
+   * <p>Supported values: - Must be >= 0 mins for serverless warehouses - Must be == 0 or >= 10 mins
+   * for non-serverless warehouses - 0 indicates no autostop.
    *
    * <p>Defaults to 120 mins
    */