+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/workspace/ListReposResponse.java
@@ -11,13 +11,13 @@ import java.util.Objects;
 @Generated
 public class ListReposResponse {
   /**
-   * Token that can be specified as a query parameter to the GET /repos endpoint to retrieve the
+   * Token that can be specified as a query parameter to the `GET /repos` endpoint to retrieve the
    * next page of results.
    */
   @JsonProperty("next_page_token")
   private String nextPageToken;
 
-  /** */
+  /** List of Git folders (repos). */
   @JsonProperty("repos")
   private Collection<RepoInfo> repos;